+++ src/d3d12/d3d12_structs.hpp
@@ -108,12 +108,12 @@ namespace wr::d3d12
 			StagingBuffer* vertex_buffer = nullptr;
 			std::optional<StagingBuffer*> index_buffer;
 
-			std::uint64_t m_num_vertices = 0;
-			std::uint64_t m_num_indices = 0;
-			std::uint64_t m_vertices_offset = 0;
-			std::uint64_t m_indices_offset = 0;
+			std::uint32_t m_num_vertices = 0;
+			std::uint32_t m_num_indices = 0;
+			std::uint32_t m_vertices_offset = 0;
+			std::uint32_t m_indices_offset = 0;
 
-			std::uint64_t m_vertex_stride = 0;
+			std::uint32_t m_vertex_stride = 0;
 		};
 
 		struct StateObjectDesc
@@ -122,9 +122,9 @@ namespace wr::d3d12
 			std::vector<std::wstring> m_library_exports;
 			std::vector<std::pair<std::wstring, std::wstring>> m_hit_groups; // first = hit group | second = entry
 
-			std::uint64_t max_payload_size;
-			std::uint64_t max_attributes_size;
-			std::uint64_t max_recursion_depth;
+			std::uint32_t max_payload_size;
+			std::uint32_t max_attributes_size;
+			std::uint32_t max_recursion_depth;
 
 			std::optional<RootSignature*> global_root_signature;
 			std::optional<std::vector<RootSignature*>> local_root_signatures;
